+++ mplsoccer/__init__.py
@@ -13,3 +13,4 @@ from .utils import *
 from .bumpy_chart import *
 from .py_pizza import *
 from .grid import *
+from .dimensions import PositionLineType, FormationHelper